+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 1.0.1
+
+- Added option for breadcrumbs
+- Fixed imports in the Admin class
+
 ## Version 1.0
 
 - Added alternate tree